Start converting website build into make
While I enjoyed hakyll for haskell and its expressiveness, it just takes
too long to build the site and requires a specific environment.

This is also coinciding with my dislike of the over complication of
build chains and tooling with software development in general.

Things should strive to be simple. Simple should be the goal until real
needs require the addition of more complexity.